sims/net/wire: fix crash when sync just unnecessary and retry on failure
The current timestamp for sims/net/wire is always the minimum of the two
attached Ethernet devices. With this, it is possible that a sync is
unnecessary in the case where eth_latency > sync_period and a message is
being forwarded from one port to another. The sync for the latter in the
next iteration will then be unnecessary.

Additionally, we now retry when a sync fails due to the shared memory
queue being full. This can be a problem when the simulator on the other
side is rather slow and sims/net/wire is producing synchronization
messages at a high rate.