lib/simbricks/nicbm: separate ethernet and PCI sync period
When simulating two hosts with a 10 ms latency link between them, `warn:
SimbricksNicIfSync failed` was printed constantly by the NICs. The
reason is that it wasn't possible to reduce the synchronization period
for the Ethernet side for the NIC since it was coupled with the PCI
side.

Now, you adjust both seperately, which allows you to realize the
described system by setting the Ethernet sync period on both the NIC and
the network to 5 ms. This allows for an efficient representation of
systems with an Ethernet latency much larger than the PCI latency since
a lot less synchronization messages have to be sent around.